fixed the lock-free lru cache, added benchmarks
The first version that kept separate the timestamps and the data
wasn't working correctly; even if the atomic set/get operations
should enforce a consistent state in the shared objects at those
checkpoints, in some scenarios the values read didn't match the
key and yet the timestamp could be successfully updated with a CAS
operation.

The solution was to group all key, value and timestamp in the
same object and update the three at once in an AtomicReferenceArray
object.

The drawback of this approach is that it puts more pressure on the
garbage collector, as new objects have to be created for each
access operation.